pyproject.toml
[project]
name = "rag-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.129.2",
    "google-genai>=1.65.0",
    "google-generativeai>=0.8.6",
    "groq>=1.0.0",
    "inngest>=0.5.15",
    "llama-index-core>=0.14.15",
    "llama-index-readers-file>=0.5.6",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.17.0",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.54.0",
    "unicorn>=2.1.4",
    "uvicorn>=0.41.0",
]
